+++ agency/execution/executor/customization_points/async_execute.hpp
@@ -6,6 +6,7 @@
 #include <agency/detail/type_traits.hpp>
 #include <agency/execution/executor/detail/utility/bulk_async_execute_with_one_shared_parameter.hpp>
 #include <agency/execution/executor/executor_traits.hpp>
+#include <agency/detail/shape_cast.hpp>
 
 
 namespace agency
