+++ agency/execution/executor/detail/utility/bulk_sync_execute_with_auto_result.hpp
@@ -5,6 +5,7 @@
 #include <agency/execution/executor/executor_traits.hpp>
 #include <agency/execution/executor/detail/utility/bulk_sync_execute_with_void_result.hpp>
 #include <agency/execution/executor/detail/utility/bulk_sync_execute_with_collected_result.hpp>
+#include <agency/container/executor_container.hpp>
 #include <agency/detail/factory.hpp>
