+++ agency/execution/executor/detail/utility/executor_container_or_void.hpp
@@ -2,7 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/execution/executor/executor_traits/is_bulk_executor.hpp>
-#include <agency/execution/executor/executor_traits/executor_container.hpp>
+#include <agency/container/executor_container.hpp>
 #include <agency/detail/type_traits.hpp>
 
 namespace agency
