+++ agency/execution/executor/executor_traits.hpp
@@ -2,7 +2,6 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/execution/executor/executor_traits/executor_allocator.hpp>
-#include <agency/execution/executor/executor_traits/executor_container.hpp>
 #include <agency/execution/executor/executor_traits/executor_execution_category.hpp>
 #include <agency/execution/executor/executor_traits/executor_execution_depth.hpp>
 #include <agency/execution/executor/executor_traits/executor_future.hpp>
